test: split object action lifecycle regressions
