+++ agent-handoff.md
@@ -11,6 +11,16 @@ existing width/LEFT preview runner. The moved bodies match the former source
 range byte-for-byte; a focused Debug build and CTest must pass before protected
 validation.
 
+## V1 visual-object action lifecycle test module
+
+The bounded #2564 structural slice moves five visual-object action lifecycle
+regressions into `test_studio_host_json_object_actions.cpp`. Existing runner
+membership, declarations, invocation order, fixtures, assertions, and Studio
+behavior remain unchanged. The extracted module is compiled only by the
+existing object-action lifecycle runner. The moved bodies match the former
+source range byte-for-byte; a focused Debug build and CTest must pass before
+protected validation.
+
 ## V1 Studio detail-header/footer section-top test module
 
 The bounded #2564 structural slice moves the four paired report/label
